Prevent memory leak of Callback in ImageStream, update README
The Callback class's Javadoc says that direct references to it should be managed by clients, because ResolveUriTask uses a WeakReference for its Callback.

Unfortunately, our README didn't accurately reflect this important part of its usage (the code snippets showed anonymous inner Callback classes), and our ImageStream implementation also was not handling it properly.

This commit updates the relevant code snippets in the README as well as the ImageStream implementation class.